Create Message Db for Api Insert and Update

+++ NewsWebsite.ViewModels/Api/Budget/BudgetArea/BudgetModal3AreaUpdateParam.cs
@@ -3,6 +3,6 @@
     public class BudgetModal3AreaUpdateParam
     {
         public int Id { get; set; }
-        public int mosavab { get; set; }
+        public long mosavab { get; set; }
     }
 }
